+++ esmerald/core/directives/operations/shell/__init__.py
@@ -1,0 +1,1 @@
+from .base import shell as shell  # noqa